+++ frontend/src/class_profesor.js
@@ -1,0 +1,17 @@
+function Clase() {
+    return (
+      <div>
+          <div
+            className="col-lg-6 d-flex flex-column justify-content-center pt-4 pt-lg-0 order-2 order-lg-1"
+            data-aos="fade-up"
+            data-aos-delay="200"
+          >
+            <h2>Clase Profesor</h2>
+            <h4>Hola!!</h4>
+          </div>
+  
+      </div>
+    );
+  }
+  
+  export default Clase;